refactor: Swap ports of runner and launcher
Change launcher's default healthcheck port from 5681 to 5680. Also
the runner's healthcheck port was changed from 5680 to 5681.
Now that launcher is responsible for runner's health, external orchestrators
need to monitor the launchers health and not the runners. By making this
change already configured healthcheck probes for the task runner
using the launcher don't have to be changed.