refactor: Simplify configuration
Remove`N8N_RUNNER_URI` and `N8N_RUNNERS_SERVER_ENABLED` env vars.
These can be statically. For the launcher to be able to monitor the runner, the
`N8N_RUNNERS_SERVER_ENABLED` env var needs to be set to `true`. Hence
there's not reason to let it be configurable. Same applies for `N8N_RUNNER_URI`,
which we know is going to be localhost, since the launcher starts the runner.